Simplify docker image build process
This is still a WIP, as we run into dependency problems caused by
Deployer 6.x not having support for PHP ^8.0, see for example
https://github.com/deployphp/deployer/pull/3189.
I think it would be wise to fork Deployer and make the 6.x branch
compatible with PHP ^8.0. When Deployer 7 is released, we can start
making use of that version for the PHP ^8.0 variant and fall back to
Deployer 6.x for PHP 7.x.

Changes that have been made:
- Install box without Composer. With this change we didn't need the
`build/` directory anymore, so it was deleted.
- Just use one PHP image for both build and runtime
- Install Node.js and NPM using nodesource rather than the Node.js Docker image
- Removed dep sentry (in the future we need to implement our own APM)
- Removed dep deployer/recipes, this packages is deprecated and we
should make use of the recipes within the deployer/deployer package.