brancher: Poll for SSH reachability after delivery
This prevents continuing the process and deploy to a brancher which we cannot reach yet. So essentially this wait for DNS to settle and potentially other things.

The SSH polling continues on the same time elapsed and timeout that the initial polling system used. So if a Brancher node is delivered in 1200 seconds, it only has 300 seconds left for SSH to become reachable (given the default 1500 seconds timeout).